requirements.txt
discord.py
discord_components
